make menu item card price darker; extract review card into separate component
