+++ src/dataTypes.tsx
@@ -5,3 +5,10 @@ export interface MenuItem {
   price: string
   description: string
 }
+
+export interface ReviewCardProps {
+  rating: string
+  name: string
+  imageUrl: string
+  review?: string
+}
