+++ tailwind.config.js
@@ -16,6 +16,7 @@ export default {
       'primary-green': '#495E57',
       'primary-green-faded': 'rgba(73,94,87,0.5)',
       'primary-yellow': '#F4CE14',
+      'dark-yellow': '#EE9972',
       'secondary-dark': '#EE9972',
       'secondary-light': '#FBDABB',
       'secondary-light-faded': 'rgba(251, 218, 187, 0.75)',
